Lots of changes
Firstly, made it so the report command must be called before another signup is called. Secondly, made it so the cancel command will actually cancel the queue instead of just the refreshing part. Thirdly, added a few commands to recalculate the average combat score since it wasn't calculating correctly, and made it so total_rounds was stored in the database.